ABOS SOTS new layer
This is a new store and new layer for the abos_sots dataset. This
is suppose to replace the old layers/harvesters. A dummy RC record
is link to this layer which will need to be modified before prod.
The harvester needs to be redeployed to RC before reviewing this
layer